pytool/pyproject.toml
[tool.poetry]
name = "pytool"
version = "0.1.121"
description = "A tool for MD simulation"
authors = ["2lu3 <hi2lu3@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
# chain
set-chain-id = "pytool.chain.set_chain_id:set_chain_id_to_command"

# directory
increment-dir = 'pytool.directory.increment_directory_index:increment_directory_index_to_command'
copy-template='pytool.directory.copy_template:copy_template_to_command'

# log_analyzer
analyze-production = 'pytool.log_analyzer.production:command'
analyze-equilibration = 'pytool.log_analyzer.equilibration:command'
analyze-minimization= 'pytool.log_analyzer.minimization:command'

# plot
plot-2d='pytool.plot.plot_2d:plot_2d_to_command'

# sequence
pickup-aligned-seq = "pytool.sequence.pickup_aligned_sequence:main"
rename2charmm = 'pytool.sequence.rename2charmm:main'

# structure
residue-distance = "pytool.structure.residue_distance:residue_distance_to_command"
atom-distance = "pytool.structure.atom_distance:atom_distance_to_command"
rmsd = "pytool.structure.rmsd:rmsd_to_command"
rmsd-trajectory = "pytool.structure.rmsd:rmsd_trajectory_to_command"
get-boxsize = 'pytool.structure.get_box_size:get_box_size_to_command'
align = 'pytool.structure.align:align_to_command'
align-trajectory='pytool.structure.align:align_trajectory_to_command'

# trajectory
concat-dcd = "pytool.trajectory.__init__:concat_dcd"
sample-trajectory='pytool.trajectory.sample_trajectory:command'


[tool.poetry.dependencies]
python = ">=3.9,<3.13"
mdanalysis = "^2.6.0"
click = "^8.1.7"
natsort = "^8.4.0"
loguru = "^0.7.0"
toml = "^0.10.2"
biopython = "^1.81"
jinja2 = "^3.1.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
